Evaluator
Need to reason about the types in the evaluator a little bit better. We really
need to *know* the types, moreso than in the type checker (where a range of
union of types could be allowable). Should the type checker help us to decipher
this information (i.e. infer the correct usage of types)?