Updated AST, parser, and type-checker
The AST now supports the mode declaration, with included state and main
declarations only.

The parser and lexer have been updated to support the new features. Note that
the grammar requires state declarations to have an initializing expression
(which needs to also be constant foldable, for some reason--I don't think it
needs to be this way, but I have decided it will be a limitation for the time
being).

The context in the type checker is starting to be updated to reflect two classes
of variables (state and local variables). It desparately needs refactoring toand
separation across multiple files.

The evaluator has barely been changed (only one because it relies on the
type-checker*). That will have to be next.

(*) In theory, what should happen is that the type-checker should turn the
program into a new intermediate representation where we specify all the types
explicitly (which will be useful for when we have type inference).