Sharing what I have to my other computer.
I am going to rewrite the context functions, since I tried to parameterize it
for its two use cases (type checker and evaluator) and it is a complete mess.
Perhaps I am going to turn it into two types or come up with an even better way
to parameterize things.