Something of a palette system.
Introducing palettes which can be compiled and sent to the SNES through DMA.

Right now, one can specify a list of 4bpp palettes that are intended for
sprites. We check if the RGB triplet correctly specifices a 15-bit
color and that the palette contains 16 colors.

Currently, the first 8 declared palettes are sent to CGRAM in
then sprites section. We should determine dynamically which palettes should be
active in CGRAM. This also means that the index of a palette in CGRAM will
always be the same.

The process of uploading the palette relies on a hardcoded label called
"palettes" where the address of the starting palette is. Since we only have one
type of palette, we just accept the data here as all the palettes, and upload
128 bytes to CGRAM in the sprites section (this means if you have fewer than 8
palettes, garbage data is also sent into CGRAM... see the color of your
program code!). This uploading process is only done once, before the user
initialization function and is a hardcoded routine within the runtime system.

This system will need significant future work...