Graphics
Specified in a file, must be sprite format 4bpp.

A hardcoded number of bytes (0x1000) are DMA'd into VRAM. I have literally no
idea how many tiles this will be. Of course it will also DMA program code that
is not necessarily graphics.

The label uses a hardcoded name ("graphics") which may be reproduced by user
code. That would surely break things (same with palettes)

No typechecking or scoping (or a way to use graphics yet!)